# services/platform-api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SyncApplyResult": {
        "additionalProperties": false,
        "properties": {
          "operationId": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Operationid",
            "type": "string"
          },
          "serverSequence": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Serversequence"
          },
          "status": {
            "$ref": "#/components/schemas/SyncApplyStatus"
          }
        },
        "required": [
          "operationId",
          "status",
          "serverSequence"
        ],
        "title": "SyncApplyResult",
        "type": "object"
      },
      "SyncApplyStatus": {
        "enum": [
          "applied",
          "stale_ignored",
          "tombstone_wins",
          "duplicate"
        ],
        "title": "SyncApplyStatus",
        "type": "string"
      },
      "SyncCursorResponse": {
        "additionalProperties": false,
        "properties": {
          "contractVersion": {
            "const": "2.0.0",
            "default": "2.0.0",
            "title": "Contractversion",
            "type": "string"
          },
          "cursor": {
            "maxLength": 2048,
            "minLength": 16,
            "title": "Cursor",
            "type": "string"
          },
          "expiresAt": {
            "format": "date-time",
            "title": "Expiresat",
            "type": "string"
          },
          "highWatermark": {
            "minimum": 0.0,
            "title": "Highwatermark",
            "type": "integer"
          },
          "issuedAt": {
            "format": "date-time",
            "title": "Issuedat",
            "type": "string"
          }
        },
        "required": [
          "cursor",
          "highWatermark",
          "issuedAt",
          "expiresAt"
        ],
        "title": "SyncCursorResponse",
        "type": "object"
      },
      "SyncEntityType": {
        "enum": [
          "scan_session",
          "capture_set",
          "capture_view",
          "analysis_run",
          "observation",
          "lesion",
          "match_decision",
          "report"
        ],
        "title": "SyncEntityType",
        "type": "string"
      },
      "SyncOperationInput": {
        "additionalProperties": false,
        "properties": {
          "contractVersion": {
            "const": "2.0.0",
            "default": "2.0.0",
            "title": "Contractversion",
            "type": "string"
          },
          "deviceId": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Deviceid",
            "type": "string"
          },
          "encryptedPayload": {
            "anyOf": [
              {
                "maxLength": 1000000,
                "minLength": 16,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Encryptedpayload"
          },
          "entityId": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Entityid",
            "type": "string"
          },
          "entityType": {
            "$ref": "#/components/schemas/SyncEntityType"
          },
          "idempotencyKey": {
            "maxLength": 256,
            "minLength": 16,
            "title": "Idempotencykey",
            "type": "string"
          },
          "occurredAt": {
            "format": "date-time",
            "title": "Occurredat",
            "type": "string"
          },
          "operation": {
            "$ref": "#/components/schemas/SyncOperationKind"
          },
          "operationId": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Operationid",
            "type": "string"
          },
          "sequence": {
            "minimum": 0.0,
            "title": "Sequence",
            "type": "integer"
          },
          "tombstone": {
            "title": "Tombstone",
            "type": "boolean"
          },
          "version": {
            "exclusiveMinimum": 0.0,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "operationId",
          "idempotencyKey",
          "deviceId",
          "entityType",
          "entityId",
          "version",
          "sequence",
          "occurredAt",
          "operation",
          "encryptedPayload",
          "tombstone"
        ],
        "title": "SyncOperationInput",
        "type": "object"
      },
      "SyncOperationKind": {
        "enum": [
          "upsert",
          "delete"
        ],
        "title": "SyncOperationKind",
        "type": "string"
      },
      "SyncOperationOutput": {
        "additionalProperties": false,
        "properties": {
          "contractVersion": {
            "const": "2.0.0",
            "default": "2.0.0",
            "title": "Contractversion",
            "type": "string"
          },
          "deviceId": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Deviceid",
            "type": "string"
          },
          "encryptedPayload": {
            "anyOf": [
              {
                "maxLength": 1000000,
                "minLength": 16,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Encryptedpayload"
          },
          "entityId": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Entityid",
            "type": "string"
          },
          "entityType": {
            "$ref": "#/components/schemas/SyncEntityType"
          },
          "idempotencyKey": {
            "maxLength": 256,
            "minLength": 16,
            "title": "Idempotencykey",
            "type": "string"
          },
          "occurredAt": {
            "format": "date-time",
            "title": "Occurredat",
            "type": "string"
          },
          "operation": {
            "$ref": "#/components/schemas/SyncOperationKind"
          },
          "operationId": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Operationid",
            "type": "string"
          },
          "sequence": {
            "minimum": 0.0,
            "title": "Sequence",
            "type": "integer"
          },
          "serverSequence": {
            "exclusiveMinimum": 0.0,
            "title": "Serversequence",
            "type": "integer"
          },
          "tombstone": {
            "title": "Tombstone",
            "type": "boolean"
          },
          "version": {
            "exclusiveMinimum": 0.0,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "operationId",
          "idempotencyKey",
          "deviceId",
          "entityType",
          "entityId",
          "version",
          "sequence",
          "occurredAt",
          "operation",
          "encryptedPayload",
          "tombstone",
          "serverSequence"
        ],
        "title": "SyncOperationOutput",
        "type": "object"
      },
      "SyncPullResponse": {
        "additionalProperties": false,
        "properties": {
          "cursor": {
            "$ref": "#/components/schemas/SyncCursorResponse"
          },
          "hasMore": {
            "title": "Hasmore",
            "type": "boolean"
          },
          "operations": {
            "items": {
              "$ref": "#/components/schemas/SyncOperationOutput"
            },
            "title": "Operations",
            "type": "array"
          }
        },
        "required": [
          "operations",
          "cursor",
          "hasMore"
        ],
        "title": "SyncPullResponse",
        "type": "object"
      },
      "SyncPushRequest": {
        "additionalProperties": false,
        "properties": {
          "operations": {
            "items": {
              "$ref": "#/components/schemas/SyncOperationInput"
            },
            "maxItems": 100,
            "minItems": 1,
            "title": "Operations",
            "type": "array"
          }
        },
        "required": [
          "operations"
        ],
        "title": "SyncPushRequest",
        "type": "object"
      },
      "SyncPushResponse": {
        "additionalProperties": false,
        "properties": {
          "cursor": {
            "$ref": "#/components/schemas/SyncCursorResponse"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/SyncApplyResult"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "results",
          "cursor"
        ],
        "title": "SyncPushResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Stoma3D Platform API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v2/sync/pull": {
      "get": {
        "operationId": "pull_sync_v2_sync_pull_get",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 2048,
                  "minLength": 16,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncPullResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Pull Sync",
        "tags": [
          "sync"
        ]
      }
    },
    "/v2/sync/push": {
      "post": {
        "operationId": "push_sync_v2_sync_push_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncPushRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncPushResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Push Sync",
        "tags": [
          "sync"
        ]
      }
    }
  }
}